Rename structural module to structure
